update pojo and entity

+++ SourceQLBanHang/src/main/java/com/qlbh/model/CongtyHome.java
@@ -1,5 +1,5 @@
 package com.qlbh.model;
-// Generated 21/09/2016 7:13:01 PM by Hibernate Tools 4.3.5.Final
+// Generated 24/09/2016 3:27:00 PM by Hibernate Tools 5.2.0.Beta1
 
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
@@ -57,7 +57,7 @@ public class CongtyHome {
 		}
 	}
 
-	public Congty findById(short id) {
+	public Congty findById(Integer id) {
 		log.debug("getting Congty instance with id: " + id);
 		try {
 			Congty instance = entityManager.find(Congty.class, id);
